+++ data/index.js
@@ -1,5 +1,11 @@
+const makeForwardsResponse = require('./make_forwards_response');
 const makeInvoice = require('./make_invoice');
 const makePaymentData = require('./make_payment_data');
 const makePaymentRequest = require('./make_payment_request');
 
-module.exports = {makeInvoice, makePaymentData, makePaymentRequest};
+module.exports = {
+  makeForwardsResponse,
+  makeInvoice,
+  makePaymentData,
+  makePaymentRequest,
+};