+++ index.js
@@ -1,3 +1,4 @@
+const {makeForwardsResponse} = require('./data');
 const {makeInvoice} = require('./data');
 const {makeInvoiceSubscription} = require('./lnd');
 const {makeLnd} = require('./lnd');
@@ -6,6 +7,7 @@ const {makePaymentData} = require('./data');
 const {makePaymentRequest} = require('./data');
 
 module.exports = {
+  makeForwardsResponse,
   makeInvoice,
   makeInvoiceSubscription,
   makeLnd,