+++ lnd/make_lnd.js
@@ -1,3 +1,4 @@
+const makeForwardsResponse = require('./../data/make_forwards_response');
 const makeInvoice = require('./../data/make_invoice');
 const makeInvoiceSubscription = require('./make_invoice_subscription');
 const makePaySubscription = require('./make_pay_subscription');
@@ -7,6 +8,7 @@ const bufAsHex = buf => buf.toString('hex');
 /** Make an LND mock object for testing
 
   {
+    getForwards: <Override Get Forwards Response Function>
     getInvoice: <Override Get Invoice Response Function>
     subscribeToInvoice: <Override Subscribe to Invoice Emitter>
     subscribeToPay: <Override Subscribe to Pay Emitter>
@@ -32,7 +34,16 @@ module.exports = overrides => {
       addInvoice: ({}, cbk) => {
         return cbk(null, makeInvoice({}));
       },
+      forwardingHistory: (args, cbk) => {
+        // Exit early when there is a custom forwards response
+        if (!!overrides.getForwards) {
+          return overrides.getForwards({offset: args.index_offset}, cbk);
+        }
+
+        return cbk(null, makeForwardsResponse({offset: args.index_offset}));
+      },
       lookupInvoice: (args, cbk) => {
+        // Exit early when there is a custom invoice response
         if (!!overrides.getInvoice) {
           return overrides.getInvoice({id: bufAsHex(args.r_hash)}, cbk);
         }
@@ -42,6 +53,7 @@ module.exports = overrides => {
     },
     invoices: {
       subscribeSingleInvoice: ({}) => {
+        // Exit early when there is a custom invoice subscription
         if (!!overrides.subscribeToInvoice) {
           return overrides.subscribeToInvoice;
         }
@@ -51,6 +63,7 @@ module.exports = overrides => {
     },
     router: {
       sendPaymentV2: args => {
+        // Exit early when there is a custom payment subscription
         if (!!overrides.subscribeToPay) {
           return overrides.subscribeToPay;
         }